# code/back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/faculty-reports/self-reflections": {
      "get": {
        "operationId": "FacultyReportsController_findReflections",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facultyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all self reflections",
        "tags": [
          "Faculty Reports & Check-ins"
        ]
      },
      "post": {
        "operationId": "FacultyReportsController_submitReflection",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitSelfReflectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit or update a self reflection",
        "tags": [
          "Faculty Reports & Check-ins"
        ]
      }
    },
    "/api/faculty-reports/action-reports": {
      "get": {
        "operationId": "FacultyReportsController_findActionReports",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cycleId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facultyId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get all action reports",
        "tags": [
          "Faculty Reports & Check-ins"
        ]
      },
      "post": {
        "operationId": "FacultyReportsController_submitActionReport",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitActionReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Submit or resubmit an action report",
        "tags": [
          "Faculty Reports & Check-ins"
        ]
      }
    },
    "/api/faculty-reports/action-reports/{id}/checkin": {
      "patch": {
        "operationId": "FacultyReportsController_reviewCheckin",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Caller role for RBAC",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Role",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "admin",
                "dean",
                "hod",
                "faculty",
                "student"
              ]
            }
          },
          {
            "in": "header",
            "name": "X-User-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCheckinDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "HOD review / finalize / request revision on action report",
        "tags": [
          "Faculty Reports & Check-ins"
        ]
      }
    },
...
  },
  "info": {
    "title": "ENDUR API",
    "description": "\n      REST API for the **ENDUR Performance Review & Feedback Management System**.\n\n      ## Role-Based Access Control\n      All protected endpoints require the `X-Role` header with one of:\n      `superuser` | `admin` | `dean` | `hod` | `faculty` | `student`\n\n      Additionally, pass `X-User-Id` and `X-User-Name` headers to attribute audit logs correctly.\n\n      ## Authentication\n      Use `POST /api/auth/login` to verify credentials. The response contains the user object\n      which you should store client-side (session) and use to populate role headers.\n      ",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "SubmitSelfReflectionDto": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "expectedRatings": {
            "type": "object"
          },
          "reflectionText": {
            "type": "string"
          }
        },
        "required": [
          "facultyId",
          "courseId",
          "cycleId",
          "expectedRatings",
          "reflectionText"
        ]
      },
      "SubmitActionReportDto": {
        "type": "object",
        "properties": {
          "facultyId": {
            "type": "string"
          },
          "courseId": {
            "type": "string"
          },
          "cycleId": {
            "type": "string"
          },
          "rootCause": {
            "type": "string"
          },
          "plannedStrategies": {
            "type": "string"
          },
          "timeline": {
            "type": "string"
          }
        },
        "required": [
          "facultyId",
          "courseId",
          "cycleId",
          "rootCause",
          "plannedStrategies",
          "timeline"
        ]
      },
      "ReviewCheckinDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "hodNotes": {
            "type": "string"
          },
          "hodOutcomes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}